feat: Gesture configuration in shortcuts config
This commit adds a Gesture abstraction, which can define gestures by the
number of fingers used, and the direction the gesture is performed in.

The abstraction is modeled after the Binding type, and the resulting
Gestures type is similar to the Shortcuts type.

This is the first step in implementing configurable touchpad gestures
(https://github.com/pop-os/cosmic-comp/pull/396)

Signed-off-by: Ryan Brue <ryanbrue.dev@gmail.com>